Separando em arquivos o feed
